# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Syncitt
# during dev, should point to your local machine
host: localhost:5095
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /api/roles:
    # binds a127 app logic to a route
    x-swagger-router-controller: roles
    get:
      tags: ["Roles"]
      description: Get roles
      summary: Get roles
      # used as the method name of the controller
      operationId: getRoles
      parameters:
        - name: query
          in: query
          description: query
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      tags: ["Roles"]
      description: Add role
      summary: Add role
      # used as the method name of the controller
      operationId: addRole
      parameters:
        - name: body
          in: body
          description: user infomation
          schema:
            $ref: "#/definitions/addRoles"
          required: true
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /roles/{id}:
    # binds a127 app logic to a route
    x-swagger-router-controller: roles
    get:
      tags: ["Roles"]
      description: Get role by id
      summary: Get role
      # used as the method name of the controller
      operationId: getRoleById
      parameters:
        - name: id
          in: path
          description: role id
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      tags: ["Roles"]
      description: update role
      summary: Update role
      # used as the method name of the controller
      operationId: updateRole
      parameters:
        - name: id
          in: path
          description: role id
          required: true
          type: string
        - name: body
          in: body
          description: Body
          required: false
          schema:
            # a pointer to a definition
            $ref: "#/definitions/addRoles"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      tags: ["Roles"]
      description: delete role
      summary: Delete role
      # used as the method name of the controller
      operationId: deleteRole
      parameters:
        - name: id
          in: path
          description: role id
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /api/getUserDefaultRoles:
    # binds a127 app logic to a route
    x-swagger-router-controller: roles
    post:
      tags: ["Roles"]
      description: Get user default roles
      summary: Get user default roles
      # used as the method name of the controller
      operationId: getUserDefaultRoles
      parameters:
        - name: body
          in: body
          description: direct login
          schema:
            $ref: "#/definitions/userRoleId"
          required: true
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/gobalSearch:
    # binds a127 app logic to a route
    x-swagger-router-controller: searchs
    post:
      tags: ["Gobal Search"]
      description: Get gobal search data
      summary: Get gobal search data
      # used as the method name of the controller
      operationId: globalSearch
      parameters:
        - name: body
          in: body
          description: text to be searched
          schema:
            $ref: "#/definitions/gobalSearchText"
          required: true
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /api/saveUserMultiRoles:
    # binds a127 app logic to a route
    x-swagger-router-controller: roles
    post:
      tags: ["Roles"]
      description: save user multiRoles
      summary: save user multiRoles
      # used as the method name of the controller
      operationId: saveUserMultiRoles
      parameters:
        - name: body
          in: body
          description: save user multiRoles
          schema:
            $ref: "#/definitions/multiRoles"
          required: true
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  addRoles:
    type: object
    properties:
      title:
        type: string
      description:
        type: string
      permission_id:
        type: array
        items:
          type: string
...
  gobalSearchText:
    type: object
    properties:
      text:
        type: string
...
  userRoleId:
    type: object
    properties:
      user_id:
        type: string
...
  multiRoles:
    type: object
    properties:
      user_id:
        type: string
      roles:
        type: array
        items:
          type: string
...
  appResponse:
    required:
      - code
    properties:
      code:
        type: number
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string